requirements.txt
matplotlib>=3.7.0
numpy>=1.24.0
pyyaml>=6.0
tabulate>=0.9.0
pytest>=7.4.0

